\name{Peak2Trough-package}
\alias{Peak2Trough-package}
\alias{Peak2Trough}
\docType{package}
\title{
Estimation of the peak-to-trough ratio of a seasonal variation component
}
\description{
Estimation of the peak-to-trough ratio of a seasonal variation component, using three different estimators. 
}
\details{
\tabular{ll}{
Package: \tab Peak2Trough\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-04-28\cr
License: \tab GPL (>= 2) \cr
}
This package is developed and used for the Monte Carlo simulation study in the paper: A. L. Christensen, S. Lundbye-Christensen and C. Dethlefsen. Poisson regression models outperform the geometrical model in estimating the peak-to-trough ratio of seasonal variation: a simulation study. Submitted, 2011. It holds seven functions: three estimators and four simulation schemes.}
\author{
Anette Luther Christensen <anluc@rn.dk>     

Maintainer: Anette Luther Christensen <anluc@rn.dk>
}
\references{
J. H. Edwards. The recognition and estimation of cyclic trends. Annals of Human Genetics 25 (1961) 83-86.

J. A. Nelder and R. W. M. Wedderburn. Generalized linear models. Journal of Royal Statistical Society 135 (1972) 370-384.

M. A. Brookhart and K. J. Rothman. Simple estimators of the intensity of seasonal occurrence. BMC Medical Research Methodology 8 (2008) 67-75.

A. L. Christensen, S. Lundbye-Christensen and C. Dethlefsen. Poisson regression models outperform the geometrical model in estimating the peak-to-trough ratio of seasonal variation: a simulation study. Submitted, 2011.
}

\keyword{ package }
