\name{Peak2Trough-package}
\alias{Peak2Trough-package}
\alias{Peak2Trough}
\docType{package}
\title{
Estimation of the peak-to-trough ratio of a seasonal variation component
}
\description{
Estimation of the peak-to-trough ratio of a seasonal variation componentbased on the geometrical model, Poisson regression and state space model.
}
\details{
\tabular{ll}{
Package: \tab Peak2Trough\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2012-03-14\cr
License: \tab GPL (>= 2) \cr
}
This package is developed and used for the Monte Carlo simulation study in the paper: A. L. Christensen, S. Lundbye-Christensen and C. Dethlefsen. Poisson regression models outperform the geometrical model in estimating the peak-to-trough ratio of seasonal variation: a simulation study. Computer Programs and Methods in Biomedicine. 104 (2011) 333-340. Further developments of the package have been performed in order to model dynamic seasonal variation using state space models.}
\author{
Anette Luther Christensen <anluc@rn.dk>     

Maintainer: Anette Luther Christensen <anluc@rn.dk>
}
\references{
J. H. Edwards. The recognition and estimation of cyclic trends. Annals of Human Genetics 25 (1961) 83-86

J. A. Nelder and R. W. M. Wedderburn. Generalized linear models. Journal of Royal Statistical Society 135 (1972) 370-384

M. A. Brookhart and K. J. Rothman. Simple estimators of the intensity of seasonal occurrence. BMC Medical Research Methodology 8 (2008) 67-75

A. L. Christensen, S. Lundbye-Christensen and C. Dethlefsen. Poisson regression models outperform the geometrical model in estimating the peak-to-trough ratio of seasonal variation: a simulation study. Computer Programs and Methods in Biomedicine. 104 (2011) 333-340

West and Harrison. Bayesian Forecasting and Dynamic Models. 2nd, Springer, 1997

Gersch and Kitagawa. The Prediction of Time Series With Trends and Seasonalities. Journal of Business & Economic Statistics. 1 (1983) 253-264

Durbin and Koopman. Time Series Analysis by STate Space Methods. 1st, Oxford University Press, 2001
}

\keyword{ Seasonal variation }
\keyword{ Geometrical model }
\keyword{ Poisson regression }
\keyword{ State space model }
